Tanyilin

+++ modules/college.go
@@ -2,7 +2,7 @@ package models
 
 type College struct {
 	//学院id
-	Id int64 `json:"id"`
+	Id uint `json:"id"`
 	//学院名称
 	College string `json:"college"`
 }
